Refactoring header into a vue.
