+++ project/frontend/src/views/HomeView.vue
@@ -1,21 +1,5 @@
 <template>
-<rux-pop-up-menu id="popup-menu-1">
-    <rux-menu-item>
-      <router-link :to="{name: 'home'}">
-        Dashboard
-      </router-link>
-    </rux-menu-item>
-    <rux-menu-item-divider></rux-menu-item-divider>
-    <rux-menu-item @click="logout">Signout</rux-menu-item>
-</rux-pop-up-menu>
-
-<rux-global-status-bar app-domain="Full Stack Demo"
-  include-icon="true"
-  menu-icons="apps">
-    <rux-clock class="right"></rux-clock>
-    <rux-icon icon="settings" label="Settings" color="#fff" @click="displayPopupMenu" aria-controls="popup-menu-1" aria-haspopup="true"></rux-icon>
-  </rux-global-status-bar>
-  
+  <TopHeader class="header"/> 
   <div class="row">
     <div class="col-md-12">
       <Dashboard />
@@ -26,16 +10,12 @@
 <script>
 // Import local controls.
 import AlertsView from "./AlertsView.vue"
+import TopHeader from "./TopHeader.vue"
 import ContactsView from "./ContactsView.vue"
 import LoginView from "./LoginView.vue"
 import Dashboard from "./Dashboard.vue"
 import RegisterView from "./RegisterView.vue"
 
-// Import Astro controls.
-import { RuxButton } from '@astrouxds/astro-web-components/dist/components/rux-button'
-import { RuxInput } from '@astrouxds/astro-web-components/dist/components/rux-input'
-import { RuxGlobalStatusBar } from '@astrouxds/astro-web-components/dist/components/rux-global-status-bar'
-import { RuxClock } from '@astrouxds/astro-web-components/dist/components/rux-global-status-bar'
 
 export default {
   name: 'HomeView',
@@ -51,32 +31,27 @@ export default {
     ContactsView,
     LoginView,
     Dashboard,
-    RegisterView
+    RegisterView,
+    TopHeader
   },
   setup() {
-    let isAuthenticated = false
 
     const getDisplayName = () => {
      username = this.user.firstname.charAt(0) + ". " + this.user.lastname
      return username
     }
     
-    const logout = () =>{
-      localStorage.clear()
-      window.location.href = '/login'
-    }
-    return { isAuthenticated, getDisplayName , logout}
+    return {  getDisplayName }
   },
 }
 </script>
 
 <style scoped>
-.right{
-  margin-left: auto;
-  margin-right: 0px;
+.header{
+  width:100%;
 }
-
 div{
   margin: 10px;
 }
+
 </style>
